refactor(skills): split builtin.rs into submodules
Split src/skills/builtin.rs (795 lines) into a directory module:

src/skills/builtin/
├── mod.rs      (~40 lines)  — module declarations + pub use re-exports
│                              + builtin_skills() + builtin_skills_with_engine()
├── file_ops.rs (~170 lines) — FileOpsSkill struct + Default + impl Skill
├── git_ops.rs  (~120 lines) — GitOpsSkill struct + impl Skill
├── search.rs   (~60 lines)  — SearchSkill struct + impl Skill
├── permission.rs(~140 lines)— PermissionSkill struct + Default + impl Skill
├── discovery.rs(~180 lines) — SkillDiscoverySkill struct + Default + impl Skill
└── tests.rs    (~220 lines) — test module

Each public API remains unchanged. All re-exports preserved via mod.rs.

Source: issue #216
Type: refactor
